This is an initial, somewhat rough version that uses lirc modes, as oppsed to a new instance of irexec, to control the Apple TV.
The downside of this approach is that it's going to be difficult if
not impossible to use a Myth menu item to go to Apple TV, as you can't
(AFAIK) send lirc a simulated remote event from another program.

The buttons in lircrc/mythtv need some more thought, I just randomly
assigned a few.