Well. It seems to work now.
It looks like the problem had something to do with the commandir
box's ability to receive a signal, and then pass it over to an
emitter. For example, when watching appletv, you'd press a remote
button, which the CIR IR receiver would pick up, and then irexec
would cause irsend to get called, which would push another signal
out via a CIR emitter. After a few button presses, this inevitably
horked the CIR box, requiring an un/replug sequence.

I was able to run a test script (remote_test.sh, included in this
commit) that *pounded* on appletv via irsend alone. It ran fine.
That's when it occurred to me that the problem happened when the
receiver was in the mix. So I plugged the old IR receiver back in,
and also, for the sake of simplicity, decided to ditch LIRC
altogether for controlling the CIR box, and use the just-released
commandir utilities instead. This way, there's no crossover
between the reciever and the emitter, other than the exec line in
a lircrc file.

So far, so good!