Things work here, using the commandIR box for both IR reception and blasting.
I moved the sleep time up from *after* a command was invoked to *between* the
time that a signal is recieved and when another signal is blasted. It seems to
solve all of the problems that I encountered with teh box freezing up and
needing an unplug/replug.

Down side is .1 seconds of sleep between reaction to button presses, up side
is a simpler setup -- there's no more need for a separate IR reciever.