Added "back to myth" function to the mythtv lircrc file.
Now, if things get out of sync -- e.g. someone uses the HDMI switch directly to flip over to AppleTV, the "Back to Myth" button will still work. Previously, it wouldn't, because it was only defined in the appletv lircrc file, and that file was only exec'ed when the "go to appletv" menu item was chosen from the myth menu.